rename: PostDetail 컴포넌트명 SubTravelogue로 변경
