+++ src/components/CreatePost/index.ts
@@ -3,6 +3,6 @@ export { default as DatePicker } from './DatePicker';
 export { default as Location } from './Location';
 export { default as PostBasic } from './PostBasic';
 export { default as PostComplete } from './PostComplete';
-export { default as PostDetail } from './PostDetail';
 export { default as StepperButton } from './StepperButton';
+export { default as SubTravelogue } from './SubTravelogue';
 export { default as Transportation } from './Transportation';
