+++ src/constants/defaultFormValue.ts
@@ -22,3 +22,11 @@ export const travelogueFormProps: UseFormProps = {
     thumbnail: null,
   },
 };
+
+export const subTravelogueFormProps: UseFormProps = {
+  defaultValues: {
+    title: '',
+    content: '',
+    addresses: [{ country: '', city: '', spot: '' }],
+  },
+};
