+++ src/pages/post/[id].tsx
@@ -1,16 +1,16 @@
-// import { useRouter } from 'next/router';
-
 import { Box } from '@mui/material';
+import { useRouter } from 'next/router';
 import { useState } from 'react';
 
-import { PostDetail, StepperButton } from '@/components/CreatePost';
+import { StepperButton, SubTravelogue } from '@/components/CreatePost';
 
-const Post = () => {
+const SubTraveloguePage = () => {
   const [steps, setSteps] = useState(0);
+  const query = useRouter().query;
 
   return (
     <Box sx={layout}>
-      <PostDetail />
+      <SubTravelogue query={query} />
       <Box sx={{ display: 'flex', justifyContent: 'space-between' }}>
         <StepperButton format='backward' steps={steps} setSteps={setSteps} />
         <StepperButton format='complete' steps={steps} setSteps={setSteps} />
@@ -19,6 +19,6 @@ const Post = () => {
   );
 };
 
-export default Post;
+export default SubTraveloguePage;
 
 const layout = { padding: '0 24px' };
